feat: d&s
