+++ src/components/Pets/PetCanvas/index.tsx
@@ -5,6 +5,7 @@ interface PetCanvasProps {
   petImage: string;
   petName: string;
   petText: string;
+  border: "horo" | "game";
   className?: string;
 }
 
@@ -15,7 +16,8 @@ const PetCanvas = (props: PetCanvasProps) => {
     const ctx = canvas?.getContext("2d");
     console.log(ctx);
     if (ctx && canvas) {
-      ctx.fillStyle = "#3D0D40";
+      const colorBorder = props.border === "game" ? "#08878E" : "#3D0D40";
+      ctx.fillStyle = colorBorder;
       ctx.fillRect(0, 0, canvas.width, canvas.height);
       ctx.fillStyle = "#FFFFFF";
       ctx.fillRect(48, 48, canvas.width - 96, canvas.height - 96);
@@ -35,7 +37,8 @@ const PetCanvas = (props: PetCanvasProps) => {
       ctx.fillStyle = "#000000";
       ctx.textBaseline = "middle";
       ctx.textAlign = "center";
-      ctx.font = "bold 80px 'Poppins'";
+      const font = props.border == "game" ? "IBM Plex Sans Thai" : "Poppins";
+      ctx.font = "bold 80px " + font;
       ctx.fillText(props.petName, canvas.width / 2, 96 + Iheight + 96);
       drawText(ctx, props.petText, {
         x: (canvas.width - 888) / 2,
@@ -48,7 +51,7 @@ const PetCanvas = (props: PetCanvasProps) => {
         vAlign: "top",
       });
     }
-  }, [props.petImage, props.petName, props.petText]);
+  }, [props.petImage, props.petName, props.petText, props.border]);
 
   return (
     <canvas
